requirement.txt
asgiref==3.7.2
Django==3.2.23
django-ckeditor==6.7.0
django-js-asset==2.2.0
Pillow==9.5.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
